fix: Search.jsx 홈 link 추가
